Database version updates will now delete old data. But it works.
Step counter is commented out, there is no step counter on virtual device and was generating a minor error when running.

ActivityMonitorService does not seem to be updating the database. The active time display does not change to show new active times.

Not sure if this is becasue the sensor monitor is not working right, or if the updates are just not being sent to the database.

I know the database is not being updated, I can check the change times in the device files and see that the database is created but the last changed times don't change, the preference file times do change. Also the db-shm (database shared memory) file updates its access time as the database is created when called by the app.

All the updating magic is supposed to happen in the ActivityMonitorService.